(PC-31176) refactor(booking): extract EndedBookingReason in a specific component
